requirements.txt
anyio==4.4.0
backoff==2.2.1
blinker==1.8.2
Brotli==1.1.0
CacheControl==0.14.0
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
exceptiongroup==1.2.1
firebase-admin==6.5.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Compress==1.14
Flask-Mail==0.10.0
Flask-Markdown==0.3
Flask-Minify==0.42
flask-squeeze==3.0.9
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
graphql-core==3.2.3
grpcio==1.64.1
grpcio-status==1.62.2
htmlmin==0.1.12
htmlminf==0.1.13
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsmin==3.0.1
lesscpy==0.15.1
Markdown==3.6
markup==0.2.2
MarkupSafe==2.1.5
mistune==3.0.2
msgpack==1.0.8
multidict==6.0.5
ply==3.11
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
